pulse api > extract general types into hivemq-edge-integration-api module
Move integration types that aren't pulse-specific into a new
hivemq-edge-integration-api gradle build that sits next to the pulse
one. Asset / AssetProviderRegistry / ExternalAssetProvider /
IncomingMessage move under com.hivemq.edge.integration.api; the
PulseMessageProcessor / PulseMessagePublisher contracts move there too,
renamed to MessageProcessor / MessagePublisher. Edge's internal
PublishIncomingMessageImpl is renamed to IncomingMessageImpl to match.

hivemq-edge-pulse-integration-api now api-depends on the new module so
existing transitive consumers keep working.
