+++ hivemq-edge-integration-api/gradle.properties
@@ -1,0 +1,4 @@
+version=2026.8-SNAPSHOT
+
+org.gradle.caching=true
+org.gradle.jvmargs=-Xmx1024M
