+++ hivemq-edge/build.gradle.kts
@@ -136,6 +136,7 @@ dependencies {
     api(libs.hivemq.extensionsdk)
     api(libs.hivemq.edge.extensionsdk)
     api(libs.hivemq.edge.adaptersdk)
+    api(libs.hivemq.edge.integration.api)
     api(libs.hivemq.edge.pulse.integration.api)
 
     // netty
