+++ hivemq-edge/settings.gradle.kts
@@ -25,6 +25,8 @@ includeBuild("../hivemq-edge-frontend") {
     name = "hivemq-edge-frontend"
 }
 
+includeBuild("../hivemq-edge-integration-api")
+
 includeBuild("../hivemq-edge-pulse-integration-api")
 
 if (file("../../hivemq-edge-extension-sdk").exists()) {
