+++ hivemq-edge/src/main/java/com/hivemq/pulse/integration/IncomingMessageImpl.java
@@ -15,19 +15,19 @@
  */
 package com.hivemq.pulse.integration;
 
-import com.hivemq.edge.pulse.integration.api.message.IncomingMessage;
+import com.hivemq.edge.integration.api.message.IncomingMessage;
 import com.hivemq.mqtt.message.mqtt5.MqttUserProperty;
 import com.hivemq.mqtt.message.publish.PUBLISH;
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
-final class PublishIncomingMessageImpl implements IncomingMessage {
+final class IncomingMessageImpl implements IncomingMessage {
 
     private static final byte @NotNull [] EMPTY_PAYLOAD = new byte[0];
 
     private final @NotNull PUBLISH delegate;
 
-    PublishIncomingMessageImpl(final @NotNull PUBLISH delegate) {
+    IncomingMessageImpl(final @NotNull PUBLISH delegate) {
         this.delegate = delegate;
     }
