+++ hivemq-edge/src/test/java/com/hivemq/configuration/entity/pulse/PulseAssetEntityTest.java
@@ -17,7 +17,7 @@ package com.hivemq.configuration.entity.pulse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.hivemq.edge.pulse.integration.api.asset.Asset;
+import com.hivemq.edge.integration.api.asset.Asset;
 import java.util.UUID;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
